requirements.txt
asgiref==3.10.0
Django==5.2.8
django-dotenv==1.4.2
django-jazzmin==3.0.1
django-querycount==0.8.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
